requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
defusedxml==0.4.1
Django==1.10.4
djangorestframework==3.5.3
djangorestframework-xml==1.3.0
enum34==1.1.6
ipython==5.1.0
ipython-genutils==0.1.0
MySQL-python==1.2.5
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
